rich content unify gradient logic
